+++ app/views/api/partner/companies/company.json.jbuilder
@@ -1,0 +1,4 @@
+json.partial!(
+  'api/partner/companies/company',
+  company: company
+)
